requirements.txt
absl-py==1.1.0
alabaster==0.7.12
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
astroid==2.11.6
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
Babel==2.10.3
backcall==0.2.0
backports.functools-lru-cache==1.6.4
bcrypt==3.2.2
beautifulsoup4==4.11.1
binaryornot==0.4.4
black==22.6.0
bleach==5.0.1
blinker==1.4
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==2.1.0
colorama==0.4.5
commonmark==0.9.1
cookiecutter==2.1.1
cryptography==37.0.2
cycler==0.11.0
dataclasses==0.8
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.5.1
docutils==0.18.1
entrypoints==0.4
fastjsonschema==2.15.3
flake8==4.0.1
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.4
importlib-resources==5.8.0
inflection==0.5.1
intervaltree==3.0.2
ipykernel==6.15.0
ipython==7.33.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isort==5.10.1
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.0.3
jinja2-time==0.2.0
joblib==1.1.0
jsonschema==4.6.1
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
keras==2.9.0
Keras-Preprocessing==1.1.2
keyring==23.6.0
kiwisolver==1.4.3
lazy-object-proxy==1.7.1
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.0
numpydoc==1.4.0
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
paramiko==2.11.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.1.2
platformdirs==2.5.1
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydeck==0.7.1
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.12.0
pylint==2.14.4
pyls-spyder==0.4.0
Pympler==1.0.1
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.4
PyQt5-sip==12.9.0
PyQtWebEngine==5.15.4
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-lsp-black==1.2.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.4.1
python-slugify==6.1.2
pytoolconfig==1.1.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==2.0.5
PyYAML==6.0
pyzmq==23.2.0
QDarkStyle==3.0.2
qstylizer==0.2.1
QtAwesome==1.1.1
qtconsole==5.3.1
QtPy==2.1.0
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.4.4
rope==1.2.0
rsa==4.8
Rtree==1.0.0
scikit-learn==1.1.1
scipy==1.8.1
semver==2.13.0
Send2Trash==1.8.0
setuptools==62.6.0
sip==6.5.1
six==1.16.0
sklearn==0.0
smmap==5.0.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.3.1
spyder-kernels==2.3.1
streamlit==1.10.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
text-unidecode==1.3
textdistance==4.3.0
threadpoolctl==3.1.0
three-merge==0.1.1
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
toolz==0.11.2
tornado==6.1
traitlets==5.3.0
typed-ast==1.5.4
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
ujson==5.3.0
Unidecode==1.3.4
urllib3==1.26.9
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
wheel==0.37.1
widgetsnbextension==3.6.1
win-inet-pton==1.1.0
wrapt==1.14.1
yapf==0.32.0
zipp==3.8.0
